Op.2 No.1: proofreading
Measure 3, beat 2, right hand: The second eighth note of this beat
should be B-natural instead of G-sharp.

Measure 13, beat 2, right hand: The sources disagree about the lower
note of the chord on the upbeat. Follow SrcB in this case. Rationale is
described in the source file.

Measure 25, beat 1, right hand: The sources disagree about the lower
note of the two chords in this beat. Follow SrcB. See the comment in the
source file for the rationale.